solved the issue on the row with id n. 107.
this row contained a \n within a string that has been recognized by BufferedReader.readLine() as a end of line. that caused a shorter row (5 columns instead of 28) that made the splitter crash.

the issue is solved by coping the BufferedReader class with the name of NewBufferedReader and adding a new method to recognize the a new line only if the last two character was '\r\n'. the new method is called newReadLine.

this trick worked fine.

the code need some cleanup but i'm hungry.