in CsvSplitter handled the end of file and changed the policy to tell the superclass the the file reached the end, now it return null.
in CsvValidator changed all the String[] and float[] to StringArray and FloatArray.

in DataCsvRow renamed all the variable to match the lowerCamelCase and modified the setters the used float[] and String[] to use FloatArray and StringArray

FloatArray is a new class to wrap a float[] into an object

in PrintClass made a new method to print an ArrayList

in the Main (ProgettoApplication) made three arrays of string that contains the data type to validate, the name of the column, ad the type class that is returned by the validation. the names into the array are written to match the name of the method that must be called. in the while we read every line of the csv and we iter all the element of the line to call the correct validator and the correct setter of DataCsvRow in a single line. at the end we just print the whole csv.

StringArray is a new class to wrap a String[] into an object