made a new package to host all spring classes.
made a controller to get all the data from a single column with relative service.

changed all the array data to ObjArray<>.

made a new validator for the country, now the country code is stored all UPPERCASE.

made a new  class to host the data from the url with description.